static analyzer fixes

+++ include/sx/array.h
@@ -110,8 +110,8 @@ static inline void* sx__sbgrowf(void* arr, int increment, int itemsize, const sx
     int dbl_cur = arr ? (sx__sbm(arr) << 1) : 0;
     int min_needed = sx_array_count(arr) + increment;
     int m = dbl_cur > min_needed ? dbl_cur : min_needed;
-    int* p = (int*)sx__realloc(alloc, arr ? sx__sbraw(arr) : 0, itemsize * m + sizeof(int) * 2, 0,
-                               file, func, line);
+    int* p = (int*)sx__realloc(alloc, arr ? sx__sbraw(arr) : 0,
+                               (size_t)itemsize * (size_t)m + sizeof(int) * 2, 0, file, func, line);
     if (p) {
         if (!arr)
             p[1] = 0;
